Plotting fixes
git-svn-id: file:///home/svn/DVA2006@5035 1e6d2d28-e1fb-0310-b8ab-81908d8c6b6a

+++ src/java/org/deltava/dao/SetNavData.java
@@ -186,11 +186,14 @@ public class SetNavData extends DAO {
 	}
 	
 	/**
-	 * Deletes a Terminal Route from the database
+	 * Deletes a Terminal Route from the database.
 	 * @param tr the TerminalRoute bean
 	 * @throws DAOException if a JDBC error occurs
 	 */
 	public void delete(TerminalRoute tr) throws DAOException {
+		if (tr == null)
+			return;
+		
 		try {
 			prepareStatementWithoutLimits("DELETE FROM common.SID_STAR WHERE (ITEMTYPE=?) AND (ICAO=?) "
 					+ "AND (NAME=?) AND (TRANSITION=?) AND (RUNWAY=?)");
